+++ run-tests.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-vendor/bin/phpunit --coverage-text --coverage-html coverage && \
-vendor/bin/phpstan analyze -l 5 src -c phpstan.neon