+++ tests/TestCase.php
@@ -7,7 +7,7 @@ namespace EventSauce\LaravelEventSauce\Tests;
 use EventSauce\LaravelEventSauce\EventSauceServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
-class TestCase extends Orchestra
+abstract class TestCase extends Orchestra
 {
     protected function getPackageProviders($app)
     {