+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-All notable changes to `laravel-eventsauce` will be documented in this file
+All notable changes to `eventsauce/laravel-eventsauce` will be documented in this file.
+
+## [Unreleased](https://github.com/laravel/dusk/compare/1.0.0...master)
 
 ## 1.0.0 - 201X-XX-XX
 
-- initial release
+- Initial release